Update Restart Servers Remotely
a Powershell GUI Script to restart Servers and Clients remotely. This Script allow you to Add Servers manually, Edit or remote Server. In Additional you have the option to ping the servers and check for availability. the result will be shown in same windows. After you submit a restart, a Authentification windows will pop up to identifiy in domain to check that you habe the correct right to restart the servers